Renamed Non-POSIX paths. This makes the bot Unix-Compatible once again!
