+++ lib/bot.js
@@ -8,6 +8,13 @@ const util = require("util");
 const {MiniGames} = require('baileys-minigames');
 const admin = require('../config/admins.json').phone+'@s.whatsapp.net';
 const store_path = path.resolve(__dirname, require('../config/bot.json').store_path);
+const fs = require('fs');
+
+fs.mkdir(store_path, {recursive: true}, (err)=>{
+    if(err){
+        console.error(err);
+    }
+});
 
 // the store maintains the data of the WA connection in memory
 // can be written out to a file & read from it
