+++ package.json
@@ -11,6 +11,7 @@
   "license": "ISC",
   "dependencies": {
     "@adiwajshing/baileys": "^4.0.0",
+    "@adiwajshing/keyed-db": "^0.2.4",
     "@hapi/boom": "^9.1.4",
     "@vitalets/google-translate-api": "^7.0.0",
     "axios": "0.21.1",
@@ -19,16 +20,17 @@
     "emoji-regex": "^10.0.0",
     "execa": "^5.1.1",
     "fluent-ffmpeg": "^2.1.2",
+    "iconv": "^3.0.0",
+    "iconv-lite": "^0.4.24",
     "jimp": "^0.16.1",
     "mime": "^1.6.0",
     "node-fetch": "^3.2.0",
     "pino": "^7.6.5",
+    "qrcode-terminal": "^0.12.0",
     "sqlite3": "^5.0.2",
     "tinycolor2": "^1.4.2",
     "wa-sticker-formatter": "^4.3.2",
-    "yargs": "^17.3.1",
-    "iconv": "^3.0.0",
-    "iconv-lite": "^0.4.24"
+    "yargs": "^17.3.1"
   },
   "devDependencies": {
     "datauri": "^4.1.0",
